+++ Client/Layers/Tactical/Objects/PointInSpace.cs
@@ -7,7 +7,7 @@ namespace Engine.Layers.Tactical.Objects
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Direction { get; set; }
+        public double Direction { get; set; }
         public int Signature { get; set; }
         public int Speed { get; set; }
         public int PositionX { get; set; }
